Move encoding utilities into willow-encoding trait
