+++ data-model/src/lib.rs
@@ -7,7 +7,7 @@
 //! - A [zero-copy](https://en.wikipedia.org/wiki/Zero-copy) implementation of Willow [paths](https://willowprotocol.org/specs/data-model/index.html#Path) and their constituent [components](https://willowprotocol.org/specs/data-model/index.html#Component).
 //! - An implementation of Willow's [entries](https://willowprotocol.org/specs/data-model/index.html#Entry).
 //! - Utilities for Willow's entry [groupings](https://willowprotocol.org/specs/grouping-entries/index.html#grouping_entries), such as [ranges](https://willowprotocol.org/specs/grouping-entries/index.html#ranges) and [areas](https://willowprotocol.org/specs/grouping-entries/index.html#areas)
-//! - Utilities for encoding and decoding, as well as implementations of all of Willow's [encodings](https://willowprotocol.org/specs/encodings/index.html#encodings).
+//! - Implementations of various [relative encodings](https://willowprotocol.org/specs/encodings/index.html#relativity).
 //!
 //! This crate **does not yet have** anything for Willow's concept of [stores](https://willowprotocol.org/specs/data-model/index.html#store). Stay tuned!
 //!
@@ -25,7 +25,6 @@
 //! - `PD` - The type used for [`PayloadDigest`](https://willowprotocol.org/specs/data-model/index.html#PayloadDigest) (willowprotocol.org), must implement the [`PayloadDigest`] trait.
 //! - `AT` - The type used for [`AuthorisationToken`](https://willowprotocol.org/specs/data-model/index.html#AuthorisationToken) (willowprotocol.org), must implement the [`AuthorisationToken`] trait.
 
-pub mod encoding;
 mod entry;
 pub use entry::*;
 pub mod grouping;
@@ -33,3 +32,4 @@ mod parameters;
 pub use parameters::*;
 mod path;
 pub use path::*;
+mod relative_encodings;
