+++ earthstar/Cargo.toml
@@ -8,6 +8,7 @@ either = "1.10.0"
 willow-data-model = { path = "../data-model" }
 arbitrary = { version = "1.0.2", features = ["derive"]}
 ufotofu = { version = "0.4.2", features = ["std"] }
+willow-encoding = { path = "../encoding" }
 
 [lints]
 workspace = true
