+++ fuzz/fuzz_targets/u32be_encoding_random.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use willow_data_model::encoding::U32BE;
+use willow_encoding::U32BE;
 use willow_fuzz::encode::encoding_random;
 
 fuzz_target!(|data: &[u8]| {
