+++ fuzz/fuzz_targets/u64be_encoding_random.rs
@@ -1,7 +1,7 @@
 #![no_main]
 
 use libfuzzer_sys::fuzz_target;
-use willow_data_model::encoding::U64BE;
+use willow_encoding::U64BE;
 use willow_fuzz::encode::encoding_random;
 
 fuzz_target!(|data: &[u8]| {
