+++ fuzz/fuzz_targets/u8be_encoding.rs
@@ -2,7 +2,7 @@
 
 use libfuzzer_sys::fuzz_target;
 use ufotofu::local_nb::consumer::TestConsumer;
-use willow_data_model::encoding::U8BE;
+use willow_encoding::U8BE;
 use willow_fuzz::encode::encoding_roundtrip;
 
 fuzz_target!(|data: (u8, TestConsumer<u8, u16, ()>)| {
