+++ meadowcap/src/mc_capability.rs
@@ -3,9 +3,8 @@ use arbitrary::Arbitrary;
 use either::Either;
 use signature::{Error as SignatureError, Signer, Verifier};
 use ufotofu::sync::consumer::IntoVec;
-use willow_data_model::{
-    encoding::sync::Encodable, grouping::Area, Entry, NamespaceId, PayloadDigest, SubspaceId,
-};
+use willow_data_model::{grouping::Area, Entry, NamespaceId, PayloadDigest, SubspaceId};
+use willow_encoding::sync::Encodable;
 
 use crate::{
     communal_capability::{CommunalCapability, NamespaceIsNotCommunalError},
@@ -292,20 +291,20 @@ pub(super) mod encoding {
     #[syncify_replace(use ufotofu::sync::{BulkConsumer, BulkProducer};)]
     use ufotofu::local_nb::{BulkConsumer, BulkProducer};
 
-    #[syncify_replace(use willow_data_model::encoding::sync::{Encodable, Decodable, RelativeDecodable, RelativeEncodable};)]
-    use willow_data_model::encoding::{Decodable, Encodable, RelativeDecodable, RelativeEncodable};
+    #[syncify_replace(use willow_encoding::sync::{Encodable, Decodable, RelativeDecodable, RelativeEncodable};)]
+    use willow_encoding::{Decodable, Encodable, RelativeDecodable, RelativeEncodable};
 
-    use willow_data_model::encoding::{
+    use willow_encoding::{
         is_bitflagged, sync::Encodable as EncodableSync, CompactWidth, DecodeError,
     };
 
-    #[syncify_replace(use willow_data_model::encoding::sync::produce_byte;)]
-    use willow_data_model::encoding::produce_byte;
+    #[syncify_replace(use willow_encoding::sync::produce_byte;)]
+    use willow_encoding::produce_byte;
 
     #[syncify_replace(
-        use willow_data_model::encoding::sync::{encode_compact_width_be, decode_compact_width_be};
+        use willow_encoding::sync::{encode_compact_width_be, decode_compact_width_be};
     )]
-    use willow_data_model::encoding::{decode_compact_width_be, encode_compact_width_be};
+    use willow_encoding::{decode_compact_width_be, encode_compact_width_be};
 
     impl<
             const MCL: usize,
